Update README.md
A sound installation that responds to stimuli and continuious evolves via user input. This input is made possible through the use of a range of different sensors. This project employs the use of HCI (Human-Computer Interaction) and several Arduinos. The software element of this project was built using two main platforms; the sensors were coding using Arduino's IDE (Intergrated Development Environment) and the audio section was built in Pure Data.